+++ src/index.js
@@ -1,18 +1,20 @@
 // @flow
 const { batch }  = require('./batch');
-const { collect }  = require('./collect');
-const { drain }  = require('./drain');
-const { produce } = require('./produce');
-const { transduce } = require('./transduce');
-const { from, of, throwError } = require('./from');
 const { chain, concatMap, map, mergeMap }  = require('./map');
+const { collect }  = require('./collect');
 const { concat }  = require('./concat');
+const { drain }  = require('./drain');
+const { from, of, throwError } = require('./from');
 const { merge }  = require('./merge');
+const { produce } = require('./produce');
 const { reduce }  = require('./reduce');
-const { subscribe }  = require('./subscribe');
-const { tap }  = require('./tap');
-const { transform }  = require('./transform');
+const { skip }  = require('./skip');
 const { StreamError } = require('./errors');
+const { subscribe }  = require('./subscribe');
+const { take }  = require('./take');
+const { tap }  = require('./tap');
+const { transduce } = require('./transduce');
+const { transform }  = require('./transform');
 
 module.exports = {
   batch,
@@ -28,8 +30,10 @@ module.exports = {
   of,
   produce,
   reduce,
+  skip,
   StreamError,
   subscribe,
+  take,
   tap,
   throwError,
   transduce,