+++ test/batch.test.js
@@ -1,6 +1,6 @@
 // @flow
 import test from 'ava';
-import { batch, collect, of } from '../src';
+import { batch, collect, from, of } from '../src';
 
 test('Batch emits arrays of the given size from a stream', (t) => {
   return of(0, 1, 2, 3, 4, 5)
@@ -28,4 +28,11 @@ test('Batch is resilient to negative count', (t) => {
     .thru(batch(-3))
     .thru(collect())
     .then((result) => t.deepEqual(result, [['a'], ['b'], ['c']]));
+});
+
+test('Batch handles errors properly', (t) => {
+  return t.throws(
+    from(new Error('An error'))
+      .thru(batch(3))
+      .thru(collect()));
 });